requirements.txt
absl-py==1.3.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
av==9.2.0
azure-core==1.26.2
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
Django==4.1.5
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
django-storages==1.13.2
django-storages-azure==1.6.8
djangorestframework==3.14.0
dnspython==2.2.1
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.3.3
google-crc32c==1.5.0
idna==3.4
isodate==0.6.1
kiwisolver==1.4.4
matplotlib==3.6.2
mediapipe==0.9.0.1
msrest==0.7.1
multidict==6.0.4
netifaces==0.11.0
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
pyee==9.0.4
pylibsrtp==0.7.1
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
xlwt==1.3.0
yarl==1.8.2
